refactor: split dialog organisms
