+++ frontends/nextjs/src/components/ui/organisms/dialogs/alert-dialog/index.ts
@@ -1,0 +1,18 @@
+export { AlertDialog } from './AlertDialog'
+export { AlertDialogTrigger } from './AlertDialogTrigger'
+export { AlertDialogContent } from './AlertDialogContent'
+export { AlertDialogHeader } from './AlertDialogHeader'
+export { AlertDialogTitle } from './AlertDialogTitle'
+export { AlertDialogDescription } from './AlertDialogDescription'
+export { AlertDialogFooter } from './AlertDialogFooter'
+export { AlertDialogCancel } from './AlertDialogCancel'
+export { AlertDialogAction } from './AlertDialogAction'
+export type { AlertDialogProps } from './AlertDialog'
+export type { AlertDialogTriggerProps } from './AlertDialogTrigger'
+export type { AlertDialogContentProps } from './AlertDialogContent'
+export type { AlertDialogHeaderProps } from './AlertDialogHeader'
+export type { AlertDialogTitleProps } from './AlertDialogTitle'
+export type { AlertDialogDescriptionProps } from './AlertDialogDescription'
+export type { AlertDialogFooterProps } from './AlertDialogFooter'
+export type { AlertDialogCancelProps } from './AlertDialogCancel'
+export type { AlertDialogActionProps } from './AlertDialogAction'
