+++ frontends/nextjs/src/components/ui/organisms/dialogs/index.ts
@@ -1,0 +1,3 @@
+export * from './command'
+export * from './alert-dialog'
+export * from './sheet'
