+++ frontends/nextjs/src/components/ui/organisms/dialogs/sheet/index.ts
@@ -1,0 +1,16 @@
+export { Sheet } from './Sheet'
+export { SheetTrigger } from './SheetTrigger'
+export { SheetContent } from './SheetContent'
+export { SheetHeader } from './SheetHeader'
+export { SheetFooter } from './SheetFooter'
+export { SheetTitle } from './SheetTitle'
+export { SheetDescription } from './SheetDescription'
+export { SheetClose } from './SheetClose'
+export type { SheetProps } from './Sheet'
+export type { SheetTriggerProps } from './SheetTrigger'
+export type { SheetContentProps } from './SheetContent'
+export type { SheetHeaderProps } from './SheetHeader'
+export type { SheetFooterProps } from './SheetFooter'
+export type { SheetTitleProps } from './SheetTitle'
+export type { SheetDescriptionProps } from './SheetDescription'
+export type { SheetCloseProps } from './SheetClose'
